Implement Contracts.IMapper in DtoMapper Mapper class
Add ProjectReference to DontPanicLabs.Ifx.Mapping.Contracts and make
Mapper explicitly implement Contracts.IMapper. All five required Map
overloads were already present as public methods, so no implementation
changes were needed.
