+++ DontPanicLabs.Ifx.Mapping.Contracts/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+using System.Resources;
+using System.Runtime.InteropServices;
+
+[assembly: CLSCompliant(true)]
+[assembly: ComVisible(false)]
+[assembly: NeutralResourcesLanguage("en")]
