+++ DontPanicLabs.Ifx.Mapping.Contracts/IRuntimeMapper.cs
@@ -1,0 +1,5 @@
+namespace DontPanicLabs.Ifx.Mapping.Contracts;
+
+public interface IRuntimeMapper : IMapperBase
+{
+}
