+++ DontPanicLabs.Ifx.Mapping.DtoMapper.Tests.DI/Integrations/ServiceLifetimeTests.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using DontPanicLabs.Ifx.Mapping.Contracts;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Shouldly;
 using Xunit;
