+++ DontPanicLabs.Ifx.Mapping.DtoMapper.Tests.DI/ScopeTests.cs
@@ -1,3 +1,4 @@
+using DontPanicLabs.Ifx.Mapping.Contracts;
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Shouldly;
 using Xunit;
