+++ DontPanicLabs.Ifx.Mapping.DtoMapper.Tests.DI/ServiceLifetimeTests.cs
@@ -1,3 +1,4 @@
+using DontPanicLabs.Ifx.Mapping.Contracts;
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
