+++ DontPanicLabs.Ifx.Mapping.DtoMapper/Mapper.cs
@@ -1,136 +1,8 @@
 namespace DontPanicLabs.Ifx.Mapping.DtoMapper;
 
-using IObjectMappingOperationOptions = IMappingOperationOptions<object, object>;
+using DontPanicLabs.Ifx.Mapping.Contracts;
+using IObjectMappingOperationOptions = DontPanicLabs.Ifx.Mapping.Contracts.IMappingOperationOptions<object, object>;
 using Factory = Func<Type, object>;
-public interface IMapperBase
-{
-    /// <summary>
-    /// Execute a mapping from the source object to a new destination object.
-    /// The source type is inferred from the source object.
-    /// </summary>
-    /// <typeparam name="TDestination">Destination type to create</typeparam>
-    /// <param name="source">Source object to map from</param>
-    /// <returns>Mapped destination object</returns>
-    TDestination Map<TDestination>(object source);
-    /// <summary>
-    /// Execute a mapping from the source object to a new destination object.
-    /// </summary>
-    /// <typeparam name="TSource">Source type to use</typeparam>
-    /// <typeparam name="TDestination">Destination type to create</typeparam>
-    /// <param name="source">Source object to map from</param>
-    /// <returns>Mapped destination object</returns>
-    TDestination Map<TSource, TDestination>(TSource source);
-    /// <summary>
-    /// Execute a mapping from the source object to the existing destination object.
-    /// </summary>
-    /// <typeparam name="TSource">Source type to use</typeparam>
-    /// <typeparam name="TDestination">Destination type</typeparam>
-    /// <param name="source">Source object to map from</param>
-    /// <param name="destination">Destination object to map into</param>
-    /// <returns>The mapped destination object</returns>
-    TDestination Map<TSource, TDestination>(TSource source, TDestination destination);
-    /// <summary>
-    /// Execute a mapping from the source object to a new destination object with explicit <see cref="System.Type"/> objects
-    /// </summary>
-    /// <param name="source">Source object to map from</param>
-    /// <param name="sourceType">Source type to use</param>
-    /// <param name="destinationType">Destination type to create</param>
-    /// <returns>Mapped destination object</returns>
-    object Map(object source, Type sourceType, Type destinationType);
-    /// <summary>
-    /// Execute a mapping from the source object to existing destination object with explicit <see cref="System.Type"/> objects
-    /// </summary>
-    /// <param name="source">Source object to map from</param>
-    /// <param name="destination">Destination object to map into</param>
-    /// <param name="sourceType">Source type to use</param>
-    /// <param name="destinationType">Destination type to use</param>
-    /// <returns>The mapped destination object</returns>
-    object Map(object source, object destination, Type sourceType, Type destinationType);
-}
-public interface IMapper : IMapperBase
-{
-    /// <summary>
-    /// Execute a mapping from the source object to a new destination object with supplied mapping options.
-    /// </summary>
-    /// <typeparam name="TDestination">Destination type to create</typeparam>
-    /// <param name="source">Source object to map from</param>
-    /// <param name="opts">Mapping options</param>
-    /// <returns>Mapped destination object</returns>
-    TDestination Map<TDestination>(object source, Action<IMappingOperationOptions<object, TDestination>> opts);
-    /// <summary>
-    /// Execute a mapping from the source object to a new destination object with supplied mapping options.
-    /// </summary>
-    /// <typeparam name="TSource">Source type to use</typeparam>
-    /// <typeparam name="TDestination">Destination type to create</typeparam>
-    /// <param name="source">Source object to map from</param>
-    /// <param name="opts">Mapping options</param>
-    /// <returns>Mapped destination object</returns>
-    TDestination Map<TSource, TDestination>(TSource source, Action<IMappingOperationOptions<TSource, TDestination>> opts);
-    /// <summary>
-    /// Execute a mapping from the source object to the existing destination object with supplied mapping options.
-    /// </summary>
-    /// <typeparam name="TSource">Source type to use</typeparam>
-    /// <typeparam name="TDestination">Destination type</typeparam>
-    /// <param name="source">Source object to map from</param>
-    /// <param name="destination">Destination object to map into</param>
-    /// <param name="opts">Mapping options</param>
-    /// <returns>The mapped destination object</returns>
-    TDestination Map<TSource, TDestination>(TSource source, TDestination destination, Action<IMappingOperationOptions<TSource, TDestination>> opts);
-    /// <summary>
-    /// Execute a mapping from the source object to a new destination object with explicit <see cref="System.Type"/> objects and supplied mapping options.
-    /// </summary>
-    /// <param name="source">Source object to map from</param>
-    /// <param name="sourceType">Source type to use</param>
-    /// <param name="destinationType">Destination type to create</param>
-    /// <param name="opts">Mapping options</param>
-    /// <returns>Mapped destination object</returns>
-    object Map(object source, Type sourceType, Type destinationType, Action<IObjectMappingOperationOptions> opts);
-    /// <summary>
-    /// Execute a mapping from the source object to existing destination object with supplied mapping options and explicit <see cref="System.Type"/> objects
-    /// </summary>
-    /// <param name="source">Source object to map from</param>
-    /// <param name="destination">Destination object to map into</param>
-    /// <param name="sourceType">Source type to use</param>
-    /// <param name="destinationType">Destination type to use</param>
-    /// <param name="opts">Mapping options</param>
-    /// <returns>The mapped destination object</returns>
-    object Map(object source, object destination, Type sourceType, Type destinationType, Action<IObjectMappingOperationOptions> opts);
-    /// <summary>
-    /// Configuration provider for performing maps
-    /// </summary>
-    IConfigurationProvider ConfigurationProvider { get; }
-    /// <summary>
-    /// Project the input queryable.
-    /// </summary>
-    /// <remarks>Projections are only calculated once and cached</remarks>
-    /// <typeparam name="TDestination">Destination type</typeparam>
-    /// <param name="source">Queryable source</param>
-    /// <param name="parameters">Optional parameter object for parameterized mapping expressions</param>
-    /// <param name="membersToExpand">Explicit members to expand</param>
-    /// <returns>Queryable result, use queryable extension methods to project and execute result</returns>
-    IQueryable<TDestination> ProjectTo<TDestination>(IQueryable source, object parameters = null, params Expression<Func<TDestination, object>>[] membersToExpand);
-    /// <summary>
-    /// Project the input queryable.
-    /// </summary>
-    /// <typeparam name="TDestination">Destination type to map to</typeparam>
-    /// <param name="source">Queryable source</param>
-    /// <param name="parameters">Optional parameter object for parameterized mapping expressions</param>
-    /// <param name="membersToExpand">Explicit members to expand</param>
-    /// <returns>Queryable result, use queryable extension methods to project and execute result</returns>
-    IQueryable<TDestination> ProjectTo<TDestination>(IQueryable source, IDictionary<string, object> parameters, params string[] membersToExpand);
-    /// <summary>
-    /// Project the input queryable.
-    /// </summary>
-    /// <param name="source">Queryable source</param>
-    /// <param name="destinationType">Destination type to map to</param>
-    /// <param name="parameters">Optional parameter object for parameterized mapping expressions</param>
-    /// <param name="membersToExpand">Explicit members to expand</param>
-    /// <returns>Queryable result, use queryable extension methods to project and execute result</returns>
-    IQueryable ProjectTo(IQueryable source, Type destinationType, IDictionary<string, object> parameters = null, params string[] membersToExpand);
-}
-public interface IRuntimeMapper : IMapperBase
-{
-}
 internal interface IInternalRuntimeMapper : IRuntimeMapper
 {
     TDestination Map<TSource, TDestination>(TSource source, TDestination destination, ResolutionContext context, Type sourceType = null, Type destinationType = null, MemberMap memberMap = null);
