+++ DontPanicLabs.Ifx.Mapping.DtoMapper/ResolutionContext.cs
@@ -1,4 +1,5 @@
 using System.Runtime.CompilerServices;
+using DontPanicLabs.Ifx.Mapping.Contracts;
 namespace DontPanicLabs.Ifx.Mapping.DtoMapper;
 /// <summary>
 /// Context information regarding resolution of a destination value
